realm: add explicit include of gasnet_coll.h
It was always the intention of the GASNet developers that clients
using the collective extensions include gasnet_coll.h explicitly.
However, due to an implementation quirk that was not actually
necessary.  That implementation quirk has been removed in the current
'develop' branch of GASNet, which breaks builds of the Realm runtime.
So, this commit adds the missing include.