make: Export symbols from linked .a files too if SHARED_OBJECTS==1
With just `-Wl,-export_dynamic` the Mac linker exports all symbols from included
.o files onto the final executable, but still skips any unreferenced symbols
from linked static libraries. This is relevant if we compile with GASNet on Mac,
in which case `librealm.dylib` expects to find the GASNet symbols in the
executable, but because those come from a static library they are not exported
by default, causing a "Symbol not found" error at runtime.