Add ability to override test args at runtime (#253)
* Fix up cuda compile args to compile SASS for everything up to the last
architecture, adding ptx for the latest (fixing an issue with r580 when
both sass and ptx is available for a removed architecture)
* Add a REALM_DEFAULT_ARGS environment and put the default test
arguments here, prepended to what is set in the test environment at
runtime. This allows us to set the default, but allow the test
environment to override it.
* This allows us to, by default, enable GPU tests with one gpu, but
override it in the test environment for more than one GPU